// src/app/core-module/enums/lookup-configuration.enum.ts
export enum RowStatusEnum {
  ACTIVE = 1,
  IN_ACTIVE = 2,
  DELETED = 3
}

export enum GenderPreferenceEnum {
  MALE = 7,
  FEMALE = 8,
  BOTH = 9
}

export enum GenderEnum {
  MALE = 10,
  FEMALE = 11,
  OTHER = 12
}

export enum InterviewTypeEnum {
  BIOMETRIC = 13,
  VISA_INTERVIEW = 14
}

export enum InterviewStatusEnum {
  APPROVED = 15,
  REJECTED = 16,
  ADMINISTRATIVE_PROCESSING = 17,
  NEED_TO_ATTEND = 18,
  BIOMETRICS_COMPLETED = 19,
  BIOMETRICS_PENDING = 20
}

export enum IntakeIdEnum {
  FALL = 21,
  SUMMER = 22,
  SPRING = 23
}